+++ cider-inspector.el
@@ -30,6 +30,7 @@
 (require 'cl-lib)
 (require 'easymenu)
 (require 'seq)
+(require 'cider-client)
 (require 'cider-eval)
 
 ;; ===================================
@@ -37,6 +38,7 @@
 ;; ===================================
 
 (defconst cider-inspector-buffer "*cider-inspect*")
+(defconst cider-inspector-value-buffer "*cider-inspect-value*")
 
 ;;; Customization
 (defgroup cider-inspector nil
@@ -140,6 +142,7 @@ Can be turned to nil once the user sees and acknowledges the feature."
     (define-key map "n" #'cider-inspector-next-inspectable-object)
     (define-key map [(shift tab)] #'cider-inspector-previous-inspectable-object)
     (define-key map "p" #'cider-inspector-previous-inspectable-object)
+    (define-key map "P" #'cider-inspector-print-current-value)
     (define-key map ":" #'cider-inspect-expr-from-inspector)
     (define-key map "f" #'forward-char)
     (define-key map "b" #'backward-char)
@@ -389,6 +392,26 @@ current-namespace."
     (cider-inspector--render-value result)
     (message "Defined current inspector value as #'%s/%s" ns var-name)))
 
+(defun cider-inspector-print-current-value ()
+  "Print the current value of the inspector."
+  (interactive)
+  (cider-ensure-connected)
+  (cider-ensure-op-supported "inspect-print-current-value")
+  (cider-popup-buffer cider-inspector-value-buffer nil 'clojure-mode 'ancillary)
+  (cider-nrepl-send-request
+   `("op" "inspect-print-current-value"
+     ,@(cider--nrepl-print-request-plist fill-column))
+   (lambda (response)
+     (with-current-buffer (get-buffer-create cider-inspector-value-buffer)
+       (let ((inhibit-read-only t))
+         (nrepl-dbind-response response (status value)
+           (cond ((member "done" status)
+                  (goto-char (point-min)))
+                 ((member "value" response)
+                  (insert value)
+                  (newline)))))))
+   (cider-current-repl)))
+
 (defun cider-inspector-tap-current-val ()
   "Sends the current Inspector current value to `tap>'."
   (interactive)
@@ -413,6 +436,7 @@ current-namespace."
       (_ (error "No object at point")))))
 
 ;; nREPL interactions
+
 (defun cider-sync-request:inspect-pop ()
   "Move one level up in the inspector stack."
   (cider-nrepl-send-sync-request `("op" "inspect-pop")